mysite/requirements.txt
asgiref==3.5.2
attrs==23.1.0
certifi==2023.5.7
Django==4.1.7
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.3
flake8==6.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sentry-sdk==1.28.1
snowballstemmer==2.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==2.0.4
